sprite-gen/requirements_no_xformers.txt
torch>=2.0.0
torchvision>=0.15.0

diffusers>=0.21.0
transformers>=4.33.0
accelerate>=0.21.0
controlnet-aux>=0.0.7

Pillow>=9.5.0

tqdm>=4.65.0
safetensors>=0.3.0

streamlit>=1.25.0